Move camera manipulator helpers out of SceneView.kt to satisfy detekt
